requirements.txt
lenskit~=0.14.1
scikit-surprise~=1.1.1
scikit-learn~=1.1.1
requests~=2.27.1
Flask~=2.1.2
pandas~=1.4.2
numpy~=1.22.4
scipy~=1.8.1
xlrd~=2.0.1